@@ -18,16 +18,17 @@
 package pulsar
 
 import (
+	"errors"
 	"fmt"
 	"math"
-	"math/big"
-	"strings"
 	"sync"
+	"sync/atomic"
 	"time"
 
-	"github.com/gogo/protobuf/proto"
+	"google.golang.org/protobuf/proto"
 
 	pb "github.com/apache/pulsar-client-go/pulsar/internal/pulsar_proto"
+	"github.com/bits-and-blooms/bitset"
 )
 
 type messageID struct {
@@ -35,6 +36,7 @@ type messageID struct {
 	entryID      int64
 	batchIdx     int32
 	partitionIdx int32
+	batchSize    int32
 }
 
 var latestMessageID = messageID{
@@ -42,6 +44,7 @@ var latestMessageID = messageID{
 	entryID:      math.MaxInt64,
 	batchIdx:     -1,
 	partitionIdx: -1,
+	batchSize:    0,
 }
 
 var earliestMessageID = messageID{
@@ -49,6 +52,7 @@ var earliestMessageID = messageID{
 	entryID:      -1,
 	batchIdx:     -1,
 	partitionIdx: -1,
+	batchSize:    0,
 }
 
 type trackingMessageID struct {
@@ -63,13 +67,26 @@ func (id trackingMessageID) Undefined() bool {
 	return id == trackingMessageID{}
 }
 
-func (id trackingMessageID) Ack() {
+func (id trackingMessageID) Ack() error {
 	if id.consumer == nil {
-		return
+		return errors.New("consumer is nil in trackingMessageID")
 	}
 	if id.ack() {
-		id.consumer.AckID(id)
+		return id.consumer.AckID(id)
 	}
+
+	return nil
+}
+
+func (id trackingMessageID) AckWithResponse() error {
+	if id.consumer == nil {
+		return errors.New("consumer is nil in trackingMessageID")
+	}
+	if id.ack() {
+		return id.consumer.AckIDWithResponse(id)
+	}
+
+	return nil
 }
 
 func (id trackingMessageID) Nack() {
@@ -91,6 +108,25 @@ func (id trackingMessageID) ack() bool {
 		return id.tracker.ack(int(id.batchIdx))
 	}
 	return true
+}
+
+func (id trackingMessageID) ackCumulative() bool {
+	if id.tracker != nil && id.batchIdx > -1 {
+		return id.tracker.ackCumulative(int(id.batchIdx))
+	}
+	return true
+}
+
+func (id trackingMessageID) prev() trackingMessageID {
+	return trackingMessageID{
+		messageID: messageID{
+			ledgerID:     id.ledgerID,
+			entryID:      id.entryID - 1,
+			partitionIdx: id.partitionIdx,
+		},
+		tracker:  id.tracker,
+		consumer: id.consumer,
+	}
 }
 
 func (id messageID) isEntryIDValid() bool {
@@ -125,6 +161,7 @@ func (id messageID) Serialize() []byte {
 		EntryId:    proto.Uint64(uint64(id.entryID)),
 		BatchIndex: proto.Int32(id.batchIdx),
 		Partition:  proto.Int32(id.partitionIdx),
+		BatchSize:  proto.Int32(id.batchSize),
 	}
 	data, _ := proto.Marshal(msgID)
 	return data
@@ -146,6 +183,10 @@ func (id messageID) PartitionIdx() int32 {
 	return id.partitionIdx
 }
 
+func (id messageID) BatchSize() int32 {
+	return id.batchSize
+}
+
 func (id messageID) String() string {
 	return fmt.Sprintf("%d:%d:%d", id.ledgerID, id.entryID, id.partitionIdx)
 }
@@ -161,20 +202,22 @@ func deserializeMessageID(data []byte) (MessageID, error) {
 		int64(msgID.GetEntryId()),
 		msgID.GetBatchIndex(),
 		msgID.GetPartition(),
+		msgID.GetBatchSize(),
 	)
 	return id, nil
 }
 
-func newMessageID(ledgerID int64, entryID int64, batchIdx int32, partitionIdx int32) MessageID {
+func newMessageID(ledgerID int64, entryID int64, batchIdx int32, partitionIdx int32, batchSize int32) MessageID {
 	return messageID{
 		ledgerID:     ledgerID,
 		entryID:      entryID,
 		batchIdx:     batchIdx,
 		partitionIdx: partitionIdx,
+		batchSize:    batchSize,
 	}
 }
 
-func newTrackingMessageID(ledgerID int64, entryID int64, batchIdx int32, partitionIdx int32,
+func newTrackingMessageID(ledgerID int64, entryID int64, batchIdx int32, partitionIdx int32, batchSize int32,
 	tracker *ackTracker) trackingMessageID {
 	return trackingMessageID{
 		messageID: messageID{
@@ -182,6 +225,7 @@ func newTrackingMessageID(ledgerID int64, entryID int64, batchIdx int32, partiti
 			entryID:      entryID,
 			batchIdx:     batchIdx,
 			partitionIdx: partitionIdx,
+			batchSize:    batchSize,
 		},
 		tracker:      tracker,
 		receivedTime: time.Now(),
@@ -196,9 +240,23 @@ func toTrackingMessageID(msgID MessageID) (trackingMessageID, bool) {
 		}, true
 	} else if mid, ok := msgID.(trackingMessageID); ok {
 		return mid, true
+	} else if cmid, ok := msgID.(chunkMessageID); ok {
+		return trackingMessageID{
+			messageID:    cmid.messageID,
+			receivedTime: cmid.receivedTime,
+			consumer:     cmid.consumer,
+		}, true
 	} else {
 		return trackingMessageID{}, false
 	}
+}
+
+func toChunkedMessageID(msgID MessageID) (chunkMessageID, bool) {
+	cid, ok := msgID.(chunkMessageID)
+	if ok {
+		return cid, true
+	}
+	return chunkMessageID{}, false
 }
 
 func timeFromUnixTimestampMillis(timestamp uint64) time.Time {
@@ -240,6 +298,8 @@ type message struct {
 	replicatedFrom      string
 	redeliveryCount     uint32
 	schema              Schema
+	schemaVersion       []byte
+	schemaInfoCache     *schemaInfoCache
 	encryptionContext   *EncryptionContext
 	index               *uint64
 	brokerPublishTime   *time.Time
@@ -290,7 +350,18 @@ func (msg *message) GetReplicatedFrom() string {
 }
 
 func (msg *message) GetSchemaValue(v interface{}) error {
+	if msg.schemaVersion != nil {
+		schema, err := msg.schemaInfoCache.Get(msg.schemaVersion)
+		if err != nil {
+			return err
+		}
+		return schema.Decode(msg.payLoad, v)
+	}
 	return msg.schema.Decode(msg.payLoad, v)
+}
+
+func (msg *message) SchemaVersion() []byte {
+	return msg.schemaVersion
 }
 
 func (msg *message) ProducerName() string {
@@ -309,14 +380,10 @@ func (msg *message) BrokerPublishTime() *time.Time {
 	return msg.brokerPublishTime
 }
 
-func newAckTracker(size int) *ackTracker {
-	var batchIDs *big.Int
-	if size <= 64 {
-		shift := uint32(64 - size)
-		setBits := ^uint64(0) >> shift
-		batchIDs = new(big.Int).SetUint64(setBits)
-	} else {
-		batchIDs, _ = new(big.Int).SetString(strings.Repeat("1", size), 2)
+func newAckTracker(size uint) *ackTracker {
+	batchIDs := bitset.New(size)
+	for i := uint(0); i < size; i++ {
+		batchIDs.Set(i)
 	}
 	return &ackTracker{
 		size:     size,
@@ -326,8 +393,9 @@ func newAckTracker(size int) *ackTracker {
 
 type ackTracker struct {
 	sync.Mutex
-	size     int
-	batchIDs *big.Int
+	size           uint
+	batchIDs       *bitset.BitSet
+	prevBatchAcked uint32
 }
 
 func (t *ackTracker) ack(batchID int) bool {
@@ -336,12 +404,84 @@ func (t *ackTracker) ack(batchID int) bool {
 	}
 	t.Lock()
 	defer t.Unlock()
-	t.batchIDs = t.batchIDs.SetBit(t.batchIDs, batchID, 0)
-	return len(t.batchIDs.Bits()) == 0
+	t.batchIDs.Clear(uint(batchID))
+	return t.batchIDs.None()
+}
+
+func (t *ackTracker) ackCumulative(batchID int) bool {
+	if batchID < 0 {
+		return true
+	}
+	t.Lock()
+	defer t.Unlock()
+	for i := 0; i <= batchID; i++ {
+		t.batchIDs.Clear(uint(i))
+	}
+	return t.batchIDs.None()
+}
+
+func (t *ackTracker) hasPrevBatchAcked() bool {
+	return atomic.LoadUint32(&t.prevBatchAcked) == 1
+}
+
+func (t *ackTracker) setPrevBatchAcked() {
+	atomic.StoreUint32(&t.prevBatchAcked, 1)
 }
 
 func (t *ackTracker) completed() bool {
 	t.Lock()
 	defer t.Unlock()
-	return len(t.batchIDs.Bits()) == 0
+	return t.batchIDs.None()
+}
+
+func (t *ackTracker) toAckSet() []int64 {
+	t.Lock()
+	defer t.Unlock()
+	if t.batchIDs.None() {
+		return nil
+	}
+	bytes := t.batchIDs.Bytes()
+	ackSet := make([]int64, len(bytes))
+	for i := 0; i < len(bytes); i++ {
+		ackSet[i] = int64(bytes[i])
+	}
+	return ackSet
+}
+
+type chunkMessageID struct {
+	messageID
+
+	firstChunkID messageID
+	receivedTime time.Time
+
+	consumer acker
+}
+
+func newChunkMessageID(firstChunkID messageID, lastChunkID messageID) chunkMessageID {
+	return chunkMessageID{
+		messageID:    lastChunkID,
+		firstChunkID: firstChunkID,
+		receivedTime: time.Now(),
+	}
+}
+
+func (id chunkMessageID) String() string {
+	return fmt.Sprintf("%s;%s", id.firstChunkID.String(), id.messageID.String())
+}
+
+func (id chunkMessageID) Serialize() []byte {
+	msgID := &pb.MessageIdData{
+		LedgerId:   proto.Uint64(uint64(id.ledgerID)),
+		EntryId:    proto.Uint64(uint64(id.entryID)),
+		BatchIndex: proto.Int32(id.batchIdx),
+		Partition:  proto.Int32(id.partitionIdx),
+		FirstChunkMessageId: &pb.MessageIdData{
+			LedgerId:   proto.Uint64(uint64(id.firstChunkID.ledgerID)),
+			EntryId:    proto.Uint64(uint64(id.firstChunkID.entryID)),
+			BatchIndex: proto.Int32(id.firstChunkID.batchIdx),
+			Partition:  proto.Int32(id.firstChunkID.partitionIdx),
+		},
+	}
+	data, _ := proto.Marshal(msgID)
+	return data
 }
