@@ -25,14 +25,14 @@ import (
 )
 
 func TestBackoff_NextMinValue(t *testing.T) {
-	backoff := &Backoff{}
+	backoff := &DefaultBackoff{}
 	delay := backoff.Next()
 	assert.GreaterOrEqual(t, int64(delay), int64(100*time.Millisecond))
 	assert.LessOrEqual(t, int64(delay), int64(120*time.Millisecond))
 }
 
 func TestBackoff_NextExponentialBackoff(t *testing.T) {
-	backoff := &Backoff{}
+	backoff := &DefaultBackoff{}
 	previousDelay := backoff.Next()
 	// the last value before capping to the max value is 51.2 s (.1, .2, .4, .8, 1.6, 3.2, 6.4, 12.8, 25.6, 51.2)
 	for previousDelay < 51*time.Second {
@@ -42,11 +42,12 @@ func TestBackoff_NextExponentialBackoff(t *testing.T) {
 		// the jitter introduces at most 20% difference so delay is less than twice the previous value
 		assert.LessOrEqual(t, int64(float64(delay)*.8), int64(2*float64(previousDelay)))
 		previousDelay = delay
+		assert.Equal(t, false, backoff.IsMaxBackoffReached())
 	}
 }
 
 func TestBackoff_NextMaxValue(t *testing.T) {
-	backoff := &Backoff{}
+	backoff := &DefaultBackoff{}
 	var delay time.Duration
 	for delay < maxBackoff {
 		delay = backoff.Next()
@@ -54,6 +55,7 @@ func TestBackoff_NextMaxValue(t *testing.T) {
 
 	cappedDelay := backoff.Next()
 	assert.GreaterOrEqual(t, int64(cappedDelay), int64(maxBackoff))
+	assert.Equal(t, true, backoff.IsMaxBackoffReached())
 	// max value is 60 seconds + 20% jitter = 72 seconds
 	assert.LessOrEqual(t, int64(cappedDelay), int64(72*time.Second))
 }
