@@ -24,7 +24,8 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/apache/pulsar-client-go/pulsar/internal/auth"
+	"github.com/apache/pulsar-client-go/pulsar/auth"
+
 	"github.com/apache/pulsar-client-go/pulsar/log"
 )
 
@@ -45,9 +46,10 @@ type connectionPool struct {
 	auth                  auth.Provider
 	maxConnectionsPerHost int32
 	roundRobinCnt         int32
-	metrics               *Metrics
+	keepAliveInterval     time.Duration
 
-	log log.Logger
+	metrics *Metrics
+	log     log.Logger
 }
 
 // NewConnectionPool init connection pool.
@@ -55,6 +57,7 @@ func NewConnectionPool(
 	tlsOptions *TLSOptions,
 	auth auth.Provider,
 	connectionTimeout time.Duration,
+	keepAliveInterval time.Duration,
 	maxConnectionsPerHost int,
 	logger log.Logger,
 	metrics *Metrics) ConnectionPool {
@@ -64,6 +67,7 @@ func NewConnectionPool(
 		auth:                  auth,
 		connectionTimeout:     connectionTimeout,
 		maxConnectionsPerHost: int32(maxConnectionsPerHost),
+		keepAliveInterval:     keepAliveInterval,
 		log:                   logger,
 		metrics:               metrics,
 	}
@@ -78,9 +82,11 @@ func (p *connectionPool) GetConnection(logicalAddr *url.URL, physicalAddr *url.U
 		p.log.Debugf("Found connection in pool key=%s logical_addr=%+v physical_addr=%+v",
 			key, conn.logicalAddr, conn.physicalAddr)
 
-		// remove stale/failed connection
+		// When the current connection is in a closed state or the broker actively notifies that the
+		// current connection is closed, we need to remove the connection object from the current
+		// connection pool and create a new connection.
 		if conn.closed() {
-			p.log.Infof("Removed connection from pool key=%s logical_addr=%+v physical_addr=%+v",
+			p.log.Debugf("Removed connection from pool key=%s logical_addr=%+v physical_addr=%+v",
 				key, conn.logicalAddr, conn.physicalAddr)
 			delete(p.connections, key)
 			conn.Close()
@@ -95,6 +101,7 @@ func (p *connectionPool) GetConnection(logicalAddr *url.URL, physicalAddr *url.U
 			tls:               p.tlsOptions,
 			connectionTimeout: p.connectionTimeout,
 			auth:              p.auth,
+			keepAliveInterval: p.keepAliveInterval,
 			logger:            p.log,
 			metrics:           p.metrics,
 		})
