@@ -35,14 +35,18 @@ type nackMockedConsumer struct {
 	lock   sync.Mutex
 }
 
-func newNackMockedConsumer() *nackMockedConsumer {
+func newNackMockedConsumer(nackBackoffPolicy NackBackoffPolicy) *nackMockedConsumer {
 	t := &nackMockedConsumer{
 		ch: make(chan messageID, 10),
 	}
 	go func() {
 		// since the client ticks at an interval of delay / 3
 		// wait another interval to ensure we get all messages
-		time.Sleep(testNackDelay + 101*time.Millisecond)
+		if nackBackoffPolicy == nil {
+			time.Sleep(testNackDelay + 101*time.Millisecond)
+		} else {
+			time.Sleep(nackBackoffPolicy.Next(1) + 101*time.Millisecond)
+		}
 		t.lock.Lock()
 		defer t.lock.Unlock()
 		t.closed = true
@@ -74,7 +78,7 @@ func (nmc *nackMockedConsumer) Wait() <-chan messageID {
 }
 
 func TestNacksTracker(t *testing.T) {
-	nmc := newNackMockedConsumer()
+	nmc := newNackMockedConsumer(nil)
 	nacks := newNegativeAcksTracker(nmc, testNackDelay, nil, log.DefaultNopLogger())
 
 	nacks.Add(messageID{
@@ -107,7 +111,7 @@ func TestNacksTracker(t *testing.T) {
 }
 
 func TestNacksWithBatchesTracker(t *testing.T) {
-	nmc := newNackMockedConsumer()
+	nmc := newNackMockedConsumer(nil)
 	nacks := newNegativeAcksTracker(nmc, testNackDelay, nil, log.DefaultNopLogger())
 
 	nacks.Add(messageID{
@@ -150,7 +154,7 @@ func TestNacksWithBatchesTracker(t *testing.T) {
 }
 
 func TestNackBackoffTracker(t *testing.T) {
-	nmc := newNackMockedConsumer()
+	nmc := newNackMockedConsumer(new(defaultNackBackoffPolicy))
 	nacks := newNegativeAcksTracker(nmc, testNackDelay, new(defaultNackBackoffPolicy), log.DefaultNopLogger())
 
 	nacks.AddMessage(new(mockMessage1))
@@ -233,6 +237,10 @@ func (msg *mockMessage1) ProducerName() string {
 	return ""
 }
 
+func (msg *mockMessage1) SchemaVersion() []byte {
+	return nil
+}
+
 func (msg *mockMessage1) GetEncryptionContext() *EncryptionContext {
 	return &EncryptionContext{}
 }
@@ -298,6 +306,10 @@ func (msg *mockMessage2) GetReplicatedFrom() string {
 }
 
 func (msg *mockMessage2) GetSchemaValue(v interface{}) error {
+	return nil
+}
+
+func (msg *mockMessage2) SchemaVersion() []byte {
 	return nil
 }
 
